mindmap/requirements.txt
Django==1.10.3
djangorestframework==3.5.3
httpie
